Gradle tasks (#10)
* Remove user prompts

* Clean up gitignore

* Change distribution for ide support

* Make wrapper executable

* Add gradle distribution

* Add bash scripts

For a future merge of launches for all operating systems, have bash and
powershell files in the same branch. These will be launched from gradle,
and new features implemented in groovy.

* Fix encoding issues after windows/unix merge

* Add cleanup and generate linux tasks

* Remove sdk files from repository

The sdk files is now generated dynamically from the currently used sdk,
and therefore doesn't need to be checked in anymore. This file will move
it's location eventually (to the gen folder), but until then, it will be
explicitly ignored.

* Add working windows launch configuration

* Clean up gradle config